Module wasm_lsp_server::metadata

Build-time metadata for the server. Build-time metadata for the server.

Constants

CFG_ENDIAN

The endianness, given by CARGO_CFG_TARGET_ENDIAN.

CFG_ENV

The toolchain-environment, given by CARGO_CFG_TARGET_ENV.

CFG_FAMILY

The OS-family, given by CARGO_CFG_TARGET_FAMILY.

CFG_OS

The operating system, given by CARGO_CFG_TARGET_OS.

CFG_POINTER_WIDTH

The pointer width, given by CARGO_CFG_TARGET_POINTER_WIDTH.

CFG_TARGET_ARCH

The target architecture, given by CARGO_CFG_TARGET_ARCH.

CI_PLATFORM

The Continuous Integration platform detected during compilation.

DEBUG

Value of DEBUG for the profile used during compilation.

FEATURES

The features that were enabled during compilation.

FEATURES_STR

The features as a comma-separated string.

GIT_COMMIT_HASH

If the crate was compiled from within a git-repository, GIT_COMMIT_HASH contains HEAD’s full commit SHA-1 hash.

GIT_DIRTY

If the repository had dirty/staged files.

GIT_HEAD_REF

If the crate was compiled from within a git-repository, GIT_HEAD_REF contains full name to the reference pointed to by HEAD (e.g.: refs/heads/master). If HEAD is detached or the branch name is not valid UTF-8 None will be stored.

GIT_VERSION

If the crate was compiled from within a git-repository, GIT_VERSION contains HEAD’s tag. The short commit id is used if HEAD is not tagged.

HOST

The host triple of the rust compiler.

NUM_JOBS

The parallelism that was specified during compilation.

OPT_LEVEL

Value of OPT_LEVEL for the profile used during compilation.

PKG_AUTHORS

A colon-separated list of authors.

PKG_DESCRIPTION

The description.

PKG_HOMEPAGE

The homepage.

PKG_LICENSE

The license.

PKG_NAME

The name of the package.

PKG_REPOSITORY

The source repository as advertised in Cargo.toml.

PKG_VERSION

The full version.

PKG_VERSION_MAJOR

The major version.

PKG_VERSION_MINOR

The minor version.

PKG_VERSION_PATCH

The patch version.

PKG_VERSION_PRE

The pre-release version.

PROFILE

release for release builds, debug for other builds.

RUSTC

The compiler that cargo resolved to use.

RUSTC_VERSION

The output of rustc -V

RUSTDOC

The documentation generator that cargo resolved to use.

RUSTDOC_VERSION

The output of rustdoc -V

TARGET

The target triple that was being compiled for.